Added tuple data structure
Create a new folder include/tuple [template based]. We can create tuples by just instantiating time as instantiate_tuple_t(int, float). But there is a problem, we can't do that it with pointers because "*" symbol isn't a good character. So we would have to create typedefs for each pointer type.